// src/app/models/textileInfo.ts
export class TextileInfo {
    productID: number;
    textileName: string;
    textileColor: string;
    textileSpecification: string;
    cost: number;
    weight: number;
    stored: string;
    remark: number;

    // constructor(productID: number,
    //     textileName: string,
    //     textileColor: string,
    //     textileSpecification: string,
    //     cost: number,
    //     weight: number,
    //     stored: string,
    //     remark: number ) {
    //         this.productID = productID


    // }
}